+++ src/main/java/com/qingqing/stock_analyse/dao/analyse/StockYiZiBanResultMapper.java
@@ -9,7 +9,7 @@ public interface StockYiZiBanResultMapper {
 
     int insertIgnore(StockYiZiBanResult record);
 
-    StockYiZiBanResult selectByPrimaryKey(Integer id);
+    StockYiZiBanResult selectByPrimaryKey(Long id);
 
     List<StockYiZiBanResult> findAllByDate(Date date);
 }
